Beanstalk now correctly calculates the delay

+++ lib/Lacuna/DB/Result/Schedule.pm
@@ -4,9 +4,18 @@ use Moose;
 use utf8;
 no warnings qw(uninitialized);
 extends 'Lacuna::DB::Result';
-use Lacuna::Util qw(format_date);
 use DateTime;
+use Scalar::Util qw(weaken);
+use Lacuna::Util qw(format_date);
+use Digest::SHA;
+use List::MoreUtils qw(uniq);
+use Email::Stuff;
+use Email::Valid;
+use UUID::Tiny ':std';
+use Lacuna::Constants qw(INFLATION);
 use Data::Dumper;
+
+extends 'Lacuna::DB::Result';
 
 __PACKAGE__->table('schedule');
 __PACKAGE__->add_columns(
@@ -23,35 +32,41 @@ __PACKAGE__->add_columns(
 after 'insert' => sub {
     my $self = shift;
 
-#    my $earliest = DateTime->now->add( hours => 2);
-#    if ($self->delivery < $earliest) {
-        # If delivery is within the next couple of hours
-        # Then put it directly on the beanstalk queue
+    if (Lacuna->config->get('beanstalk')) {
+        # an enhancement would to only put entries on beanstalk that are due within the hour
+        # and also have an hourly cron job for entries that became due in the following hour
         $self->queue_for_delivery;
-#    }
+    }
     return $self;
 };
 
 before 'delete' => sub {
     my $self = shift;
-
-    my $queue = Lacuna->queue;
-    # Delete the job off the queue
-    $queue->delete($self->job_id);
+   
+    if (Lacuna->config->get('beanstalk')) {
+        my $queue = Lacuna->queue;
+        # Delete the job off the queue
+        $queue->delete($self->job_id);
+    }
 };
+
 
 # Put this entry onto the beanstalk queue
 #
 sub queue_for_delivery {
     my ($self) = @_;
 
-    my $dur     = $self->delivery->subtract_datetime_absolute(DateTime->now);
-    my $delay   = int($dur->in_units('seconds'));
-    $delay      = 0 if $delay < 0;
+    my $now_epoch   = DateTime->now->epoch;
+    my $delivery    = $self->delivery;
+    my $del_epoch   = $delivery->epoch;
+    my $delay       = $del_epoch - $now_epoch;
+    $delay = 0 if $delay < 0;
+    my $now         = DateTime->now;
+
+    print STDERR "DELIVERY: now: $now delivery: $delivery delay: $delay table: ".$self->parent_table." task: ".$self->task."\n";
 
     my $queue   = Lacuna->queue || 'default';
     my $priority    = $self->priority || 1000;
-#print STDERR "DELAY: $delay table: ".$self->parent_table." task: ".$self->task."\n";
     my $job = $queue->publish($self->queue,
         {
             id              => $self->id,
